refactor(be): DeliveryRepository 수정, SettlementSummaryProjection 수정

+++ backend/src/main/java/com/deliveranything/domain/settlement/dto/projection/SettlementSummaryProjection.java
@@ -5,8 +5,8 @@ import java.math.BigDecimal;
 public record SettlementSummaryProjection(
     Long totalTransactionCount,
     Long weeklyTransactionCount,
-    Long weeklySettledAmount,
-    Long monthlySettledAmount,
+    BigDecimal weeklySettledAmount,
+    BigDecimal monthlySettledAmount,
     BigDecimal totalSettledAmount
 ) {
